+++ include/hs_routes.h
@@ -5,6 +5,7 @@
 #include "hs_routes_auth.h"
 #include "hs_routes_common.h"
 #include "hs_routes_error.h"
+#include "hs_routes_favicon.h"
 #include "hs_routes_fs.h"
 #include "hs_routes_payload_limit.h"
 #include "hs_routes_powered_by.h"