+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ### v0.4.0
 
+* Fix: Incorrect handling of binary files
 * Enhancement: New options for the hs_routes_fs_file_route_new_with_options function
 * Enhancement: New hs_routes_fs_directory_route_new_with_media_support function
 * Enhancement: Added additional mime types
+* Enhancement: New favicon route
 
 ### v0.3.0 (2021-09-09)